# app/controllers/user.go
package controllers

import (
	"errors"
	"path/filepath"
	"strings"
	"time"

	"gin-fast/app/global/app"
	"gin-fast/app/models"
	"gin-fast/app/service"
	"gin-fast/app/utils/common"
	"gin-fast/app/utils/gormhelper"
	"gin-fast/app/utils/passwordhelper"
	"gin-fast/app/utils/tenanthelper"

	"strconv"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
	"gorm.io/gorm"
)

// UserController 用户控制器
// @Summary 用户管理API
// @Description 用户管理相关接口
// @Tags 用户管理
// @Accept json
// @Produce json
// @Router /users [get]
type UserController struct {
	Common
	UserService   *service.User
	CasbinService *service.PermissionService
}

// NewUserController 创建用户控制器
func NewUserController() *UserController {
	return &UserController{
		Common:        Common{},
		UserService:   service.NewUserService(),
		CasbinService: service.NewPermissionService(),
	}
}

// GetProfile 获取当前登录用户信息
// @Summary 获取当前登录用户信息
// @Description 获取当前登录用户的信息，包括角色和权限
// @Tags 用户管理
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{} "成功返回用户信息"
// @Failure 401 {object} map[string]interface{} "用户未登录"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/profile [get]
// @Security ApiKeyAuth
func (uc *UserController) GetProfile(c *gin.Context) {
	// 从上下文中获取用户ID
	claims := common.GetClaims(c)
	if claims == nil {
		uc.FailAndAbort(c, "用户未登录", errors.New("用户未登录"))
		return
	}

	user, err := uc.UserService.GetUserProfile(c, claims.UserID)
	if err != nil {
		uc.FailAndAbort(c, "获取用户信息失败", err)
	}

	// 通过 claims.TenantID 查询租户信息
	tenant := models.NewTenant()
	tenantName := ""
	tenantDomain := ""
	if claims.TenantID > 0 {
		if err := tenant.FindByID(c, claims.TenantID); err == nil && !tenant.IsEmpty() {
			tenantName = tenant.Name
			tenantDomain = tenant.Domain
		}
	}

	userTenantList := models.NewSysUserTenantList()
	tenantList := models.NewTenantList() // 关联的租户列表
	if user.TenantID > 0 {
		// 获取关联的租户信息
		if err := userTenantList.Find(c, func(d *gorm.DB) *gorm.DB {
			return d.Where("user_id = ?", claims.UserID).Preload("Tenant")
		}); err != nil {
			uc.FailAndAbort(c, "查询用户关联租户错误", err)
		}
		tenantList = userTenantList.GetTenants()
	} else {
		// 如果用户默认租户ID，则获取所有关联的租户
		if err := tenantList.Find(c); err != nil {
			uc.FailAndAbort(c, "查询用户关联租户错误", err)
		}
	}

	defaultUserTenant := userTenantList.Filter(func(ut *models.SysUserTenant) bool {
		return ut.IsDefault == true
	})
	// 默认租户, TenantID为0时无默认租户
	var defaultTenant *models.Tenant
	if defaultUserTenant != nil {
		defaultTenant = defaultUserTenant.Tenant
	}

	uc.Success(c, gin.H{
		"id":            user.ID,
		"avatar":        user.Avatar,
		"userName":      user.Username,
		"nickName":      user.NickName,
		"roleIDs":       user.Roles.GetRoleIDs(),
		"permissions":   user.Permissions,
		"sex":           user.Sex,
		"status":        user.Status,
		"email":         user.Email,
		"phone":         user.Phone,
		"createdAt":     user.CreatedAt,
		"description":   user.Description,
		"roles":         user.Roles,
		"department":    user.Department,
		"tenantID":      claims.TenantID, // 用户当前登录的租户的ID
		"tenantCode":    claims.TenantCode,
		"tenantName":    tenantName,
		"tenantDomain":  tenantDomain,  // 完整的域名
		"defaultTenant": defaultTenant, // 默认租户
		"tenants":       tenantList,    // 关联的租户列表
	})
}

// List 用户列表
// @Summary 用户列表
// @Description 获取用户列表，支持分页和过滤
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param pageNum query int false "页码" default(1)
// @Param pageSize query int false "每页数量" default(10)
// @Param name query string false "用户名或昵称"
// @Param phone query string false "手机号"
// @Param status query string false "状态"
// @Success 200 {object} map[string]interface{} "成功返回用户列表"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/list [get]
// @Security ApiKeyAuth
func (uc *UserController) List(c *gin.Context) {
	var req models.UserListRequest
	if err := req.Validate(c); err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}

	userList := models.NewUserList()
	total, err := userList.GetTotal(c, req.Handle(), tenanthelper.TenantScope(c))
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	err = userList.Find(c, tenanthelper.TenantScope(c), req.Paginate(), req.Handle(), func(d *gorm.DB) *gorm.DB {
		return d.Omit("password").Preload("Roles").Preload("Department")
	})
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	uc.Success(c, gin.H{
		"list":  userList,
		"total": total,
	})

}

// GetUserByID 根据ID获取用户
// @Summary 根据ID获取用户
// @Description 根据用户ID获取用户详细信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param id path int true "用户ID"
// @Success 200 {object} map[string]interface{} "成功返回用户信息"
// @Failure 400 {object} map[string]interface{} "用户ID格式错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/{id} [get]
// @Security ApiKeyAuth
func (uc *UserController) GetUserByID(c *gin.Context) {
	// 获取路径参数
	idStr := c.Param("id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		uc.FailAndAbort(c, "Invalid user ID", err)
	}

	// 获取用户信息（限定当前租户，防止跨租户读取）
	user, err := uc.UserService.GetUserProfile(c, uint(id), tenanthelper.TenantScope(c))
	if err != nil {
		uc.FailAndAbort(c, "获取用户信息失败", err)
	}
	user.Password = ""
	uc.Success(c, user)
}

// Add 新增用户
// @Summary 新增用户
// @Description 创建新用户
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param user body models.AddRequest true "用户信息"
// @Success 200 {object} map[string]interface{} "用户创建成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/add [post]
// @Security ApiKeyAuth
func (uc *UserController) Add(c *gin.Context) {
	var req models.AddRequest
	if err := req.Validate(c); err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}

	// 校验角色必须属于当前租户，防止跨租户授权提权（全局上下文租户为0时不校验）
	if currentTenantID := common.GetCurrentTenantID(c); currentTenantID > 0 && len(req.Roles) > 0 {
		roleList := models.NewSysRoleList()
		err := roleList.Find(c, func(db *gorm.DB) *gorm.DB {
			return db.Where("id IN ? AND tenant_id = ?", req.Roles, currentTenantID)
		})
		if err != nil {
			uc.FailAndAbort(c, "验证角色信息失败", err)
		}
		if len(roleList) != len(req.Roles) {
			uc.FailAndAbort(c, "存在不属于当前租户的角色", nil)
		}
	}

	// 检查用户名是否已存在
	user := models.NewUser()
	err := user.GetUserByUsername(c, req.UserName)
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	if !user.IsEmpty() {
		uc.FailAndAbort(c, "用户已存在", nil)
	}

	// 检查手机号是否已被其他用户使用
	if req.Phone != "" {
		existUser := models.NewUser()
		err = existUser.GetUserByPhone(c, req.Phone)
		if err != nil {
			uc.FailAndAbort(c, err.Error(), err)
		}
		if !existUser.IsEmpty() {
			uc.FailAndAbort(c, "手机号已被其他用户使用", nil)
		}
	}

	// 检查邮箱是否已被其他用户使用
	if req.Email != "" {
		existUser := models.NewUser()
		err = existUser.GetUserByEmail(c, req.Email)
		if err != nil {
			uc.FailAndAbort(c, err.Error(), err)
		}
		if !existUser.IsEmpty() {
			uc.FailAndAbort(c, "邮箱已被其他用户使用", nil)
		}
	}

	// 密码加密
	hashedPassword, err := passwordhelper.HashPassword(req.Password)
	if err != nil {
		uc.FailAndAbort(c, "Failed to hash password", err)
	}

	// 使用事务创建用户、角色关联与casbin策略（任一步失败整体回滚）
	err = uc.CasbinService.RunWithCasbin(c, func(tx *gorm.DB, ops app.CasbinInterf) error {
		// 创建用户
		user.Username = req.UserName
		user.NickName = req.NickName
		user.Phone = req.Phone
		user.Email = req.Email
		user.Password = string(hashedPassword)
		user.Sex = req.Sex
		user.DeptID = req.DeptId
		user.Status = req.Status
		user.Description = req.Description

		if err := tx.Create(user).Error; err != nil {
			return err
		}

		// 创建用户角色关联
		if len(req.Roles) > 0 {
			userRoles := make([]models.SysUserRole, len(req.Roles))
			for i, roleID := range req.Roles {
				userRoles[i] = models.SysUserRole{
					UserID: user.ID,
					RoleID: roleID,
				}
			}
			if err := tx.Create(&userRoles).Error; err != nil {
				return err
			}
		}

		// 写入用户租户关联表
		if currentTenantID := common.GetCurrentTenantID(c); currentTenantID > 0 {
			err = tx.Create(&models.SysUserTenant{
				UserID:    user.ID,
				TenantID:  currentTenantID,
				IsDefault: true,
				CreatedAt: time.Now(),
			}).Error
			if err != nil {
				return err
			}
		}

		// 写入casbin用户角色关系（与业务表写同一事务）
		return uc.CasbinService.Use(ops).AddRoleForUser(c, user.ID, req.Roles)
	})

	if err != nil {
		// 前置查重后仍可能因并发提交触发唯一索引冲突（用户名全局唯一），转友好提示避免 500
		if gormhelper.IsDuplicateKeyError(err) {
			uc.FailAndAbort(c, "保存失败：用户名、手机号或邮箱已被使用", err)
		}
		uc.FailAndAbort(c, "Failed to create user", err)
	}

	uc.SuccessWithMessage(c, "User created successfully", nil)
}

// UpdateProfile 更新用户资料
// @Summary 更新用户资料
// @Description 更新用户信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param user body models.UpdateRequest true "用户信息"
// @Success 200 {object} map[string]interface{} "用户更新成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/edit [put]
// @Security ApiKeyAuth
func (uc *UserController) Update(c *gin.Context) {
	var req models.UpdateRequest
	if err := req.Validate(c); err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}

	// 校验角色必须属于当前租户，防止跨租户授权提权（全局上下文租户为0时不校验）
	if currentTenantID := common.GetCurrentTenantID(c); currentTenantID > 0 && len(req.Roles) > 0 {
		roleList := models.NewSysRoleList()
		err := roleList.Find(c, func(db *gorm.DB) *gorm.DB {
			return db.Where("id IN ? AND tenant_id = ?", req.Roles, currentTenantID)
		})
		if err != nil {
			uc.FailAndAbort(c, "验证角色信息失败", err)
		}
		if len(roleList) != len(req.Roles) {
			uc.FailAndAbort(c, "存在不属于当前租户的角色", nil)
		}
	}

	// 检查用户是否存在（限定当前租户，防止跨租户改/删）
	user := models.NewUser()
	err := user.Find(c, tenanthelper.TenantScope(c), func(db *gorm.DB) *gorm.DB {
		return db.Where("id = ?", req.Id)
	})
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	if user.IsEmpty() {
		uc.FailAndAbort(c, "用户不存在", nil, 404)
	}

	// 检查用户名是否与其他用户冲突（排除当前用户）
	existUser := models.NewUser()
	err = existUser.GetUserByUsername(c, req.UserName)
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	if !existUser.IsEmpty() && existUser.ID != req.Id {
		uc.FailAndAbort(c, "用户名已被其他用户使用", nil)
	}

	// 检查手机号是否已被其他用户使用
	if req.Phone != "" {
		existUser := models.NewUser()
		err = existUser.GetUserByPhone(c, req.Phone)
		if err != nil {
			uc.FailAndAbort(c, err.Error(), err)
		}
		if !existUser.IsEmpty() && existUser.ID != req.Id {
			uc.FailAndAbort(c, "手机号已被其他用户使用", nil)
		}
	}

	// 检查邮箱是否已被其他用户使用
	if req.Email != "" {
		existUser := models.NewUser()
		err = existUser.GetUserByEmail(c, req.Email)
		if err != nil {
			uc.FailAndAbort(c, err.Error(), err)
		}
		if !existUser.IsEmpty() && existUser.ID != req.Id {
			uc.FailAndAbort(c, "邮箱已被其他用户使用", nil)
		}
	}

	// 使用事务更新用户、角色关联与casbin策略（任一步失败整体回滚）
	err = uc.CasbinService.RunWithCasbin(c, func(tx *gorm.DB, ops app.CasbinInterf) error {
		// 更新用户信息
		user.Username = req.UserName
		user.NickName = req.NickName
		user.Phone = req.Phone
		user.Email = req.Email
		user.Sex = req.Sex
		user.DeptID = req.DeptId
		user.Status = req.Status
		user.Description = req.Description

		// 如果提供了新密码，则加密并更新密码
		if req.Password != "" {
			// 加密新密码
			hashedPassword, err := passwordhelper.HashPassword(req.Password)
			if err != nil {
				return err
			}
			user.Password = string(hashedPassword)
		}

		if err := tx.Save(user).Error; err != nil {
			return err
		}

		// 删除现有的用户角色关联
		if err := tx.Where("user_id = ?", user.ID).Delete(&models.SysUserRole{}).Error; err != nil {
			return err
		}

		// 创建新的用户角色关联
		if len(req.Roles) > 0 {
			userRoles := make([]models.SysUserRole, len(req.Roles))
			for i, roleID := range req.Roles {
				userRoles[i] = models.SysUserRole{
					UserID: user.ID,
					RoleID: roleID,
				}
			}
			if err := tx.Create(&userRoles).Error; err != nil {
				return err
			}
		}

		// 同步casbin用户角色关系（与业务表写同一事务）
		return uc.CasbinService.Use(ops).EditUserRoles(c, user.ID, req.Roles)
	})

	if err != nil {
		// 并发提交触发唯一索引冲突（用户名全局唯一）时转友好提示
		if gormhelper.IsDuplicateKeyError(err) {
			uc.FailAndAbort(c, "保存失败：用户名、手机号或邮箱已被使用", err)
		}
		uc.FailAndAbort(c, "更新用户失败", err)
	}

	uc.SuccessWithMessage(c, "更新成功", nil)
}

// Delete 删除用户
// @Summary 删除用户
// @Description 删除用户信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param user body models.DeleteRequest true "用户删除请求参数"
// @Success 200 {object} map[string]interface{} "用户删除成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/delete [delete]
// @Security ApiKeyAuth
func (uc *UserController) Delete(c *gin.Context) {
	var req models.DeleteRequest
	if err := req.Validate(c); err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}

	// 检查用户是否存在（限定当前租户，防止跨租户改/删）
	user := models.NewUser()
	err := user.Find(c, tenanthelper.TenantScope(c), func(db *gorm.DB) *gorm.DB {
		return db.Where("id = ?", req.Id)
	})
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	if user.IsEmpty() {
		uc.FailAndAbort(c, "用户不存在", nil, 404)
	}

	// 使用事务删除用户、角色关联与casbin策略（任一步失败整体回滚）
	err = uc.CasbinService.RunWithCasbin(c, func(tx *gorm.DB, ops app.CasbinInterf) error {
		// 删除用户角色关联
		if err := tx.Where("user_id = ?", user.ID).Delete(&models.SysUserRole{}).Error; err != nil {
			return err
		}
		// 删除用户租户关联
		if err := tx.Where("user_id = ?", user.ID).Delete(&models.SysUserTenant{}).Error; err != nil {
			return err
		}
		// 软删除用户
		if err := tx.Where("id = ?", user.ID).Delete(user).Error; err != nil {
			return err
		}

		// 移除casbin用户角色关系（与业务表写同一事务）
		return uc.CasbinService.Use(ops).DeleteUserRoles(c, user.ID, nil)
	})

	if err != nil {
		uc.FailAndAbort(c, "删除用户失败", err)
	}
	uc.SuccessWithMessage(c, "删除成功", nil)
}

// UpdateAccount 更新用户账户信息（密码、手机号、邮箱）
// @Summary 更新用户账户信息
// @Description 更新用户密码、手机号及邮箱信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param user body models.UpdateAccountRequest true "用户账户信息"
// @Success 200 {object} map[string]interface{} "用户账户信息更新成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/updateAccount [put]
// @Security ApiKeyAuth
func (uc *UserController) UpdateAccount(c *gin.Context) {
	var req models.UpdateAccountRequest
	if err := req.Validate(c); err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	// 该api只能更新当前用户的账户信息
	currentUserID := common.GetCurrentUserID(c)
	// 检查用户是否存在
	user := models.NewUser()
	err := user.GetUserByID(c, currentUserID)
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	if user.IsEmpty() {
		uc.FailAndAbort(c, "用户不存在", nil, 404)
	}

	// 检查手机号是否已被其他用户使用
	if req.Phone != "" {
		existUser := models.NewUser()
		err = existUser.GetUserByPhone(c, req.Phone)
		if err != nil {
			uc.FailAndAbort(c, err.Error(), err)
		}
		if !existUser.IsEmpty() && existUser.ID != currentUserID {
			uc.FailAndAbort(c, "手机号已被其他用户使用", nil)
		}
	}

	// 检查邮箱是否已被其他用户使用
	if req.Email != "" {
		existUser := models.NewUser()
		err = existUser.GetUserByEmail(c, req.Email)
		if err != nil {
			uc.FailAndAbort(c, err.Error(), err)
		}
		if !existUser.IsEmpty() && existUser.ID != currentUserID {
			uc.FailAndAbort(c, "邮箱已被其他用户使用", nil)
		}
	}
	passwordChanged := false
	if req.Password != "" {
		// 修改密码必须验证旧密码，防止token泄露后被静默改密接管账号
		if req.OldPassword == "" {
			uc.FailAndAbort(c, "请输入旧密码", nil)
		}
		if err := passwordhelper.ComparePassword(user.Password, req.OldPassword); err != nil {
			uc.FailAndAbort(c, "旧密码不正确", nil)
		}

		// 加密新密码
		hashedPassword, err := passwordhelper.HashPassword(req.Password)
		if err != nil {
			uc.FailAndAbort(c, "密码加密失败", err)
		}

		// 更新用户信息
		user.Password = string(hashedPassword)
		passwordChanged = true
	}

	if req.Phone != "" {
		user.Phone = req.Phone
	}
	if req.Email != "" {
		user.Email = req.Email
	}

	if err := app.DB().WithContext(c).Save(user).Error; err != nil {
		// 该接口仅修改手机号/邮箱，冲突提示文案相应收窄
		if gormhelper.IsDuplicateKeyError(err) {
			uc.FailAndAbort(c, "保存失败：手机号或邮箱已被使用", err)
		}
		uc.FailAndAbort(c, "更新用户信息失败", err)
	}

	// 密码修改成功后吊销当前token；refresh token按用户吊销（每用户单key），
	// 其他设备access token在剩余有效期内自然过期、无法续期
	if passwordChanged {
		if tokenString, err := common.GetAccessTokenFromHeader(c); err == nil && tokenString != "" {
			app.TokenService.RevokeTokenWithCache(tokenString)
		}
		if err := app.TokenService.RevokeRefreshToken(currentUserID); err != nil {
			app.ZapLog.Error("改密后吊销refresh token失败", zap.Error(err))
		}
		uc.SuccessWithMessage(c, "密码修改成功，请重新登录", nil)
		return
	}

	uc.SuccessWithMessage(c, "账户信息更新成功", nil)
}

// avatarImageExts 头像允许的图片扩展名白名单（小写带点）
var avatarImageExts = []string{".jpg", ".jpeg", ".png", ".gif", ".bmp"}

// isAllowedImageExt 判断文件名扩展名是否在头像图片白名单内（大小写不敏感）
func isAllowedImageExt(filename string) bool {
	ext := strings.ToLower(filepath.Ext(filename))
	for _, allow := range avatarImageExts {
		if ext == allow {
			return true
		}
	}
	return false
}

// UploadAvatar 上传用户头像
// @Summary 上传用户头像
// @Description 上传用户头像文件
// @Tags 用户管理
// @Accept multipart/form-data
// @Produce json
// @Param file formData file true "用户头像文件"
// @Success 200 {object} map[string]interface{} "用户头像上传成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/uploadAvatar [post]
// @Security ApiKeyAuth
func (uc *UserController) UploadAvatar(c *gin.Context) {
	// 从上下文中获取用户ID
	claims := common.GetClaims(c)
	if claims == nil {
		uc.FailAndAbort(c, "用户未登录", nil)
		return
	}

	// 落盘前先校验扩展名，非图片直接拒绝（零落盘），避免"先保存后拒绝"产生残留文件
	fileHeader, err := c.FormFile("file")
	if err != nil {
		uc.FailAndAbort(c, "获取上传文件失败", err)
	}
	if !isAllowedImageExt(fileHeader.Filename) {
		uc.FailAndAbort(c, "只允许上传图片文件", nil)
	}

	// 处理文件上传
	response, err := app.UploadService.HandleUpload(c, "file")
	if err != nil {
		uc.FailAndAbort(c, "文件上传失败", err)
	}

	// 兜底校验：上传服务实际保存的类型若非图片（按全局白名单放行了本接口不允许的类型），
	// 拒绝的同时删除已落盘文件，确保不留残留
	if !isAllowedImageExt(response.FileType) {
		if delErr := app.UploadService.DeleteFile(response.Url); delErr != nil {
			app.ZapLog.Error("清理非图片头像残留文件失败", zap.Error(delErr), zap.String("url", response.Url))
		}
		uc.FailAndAbort(c, "只允许上传图片文件", nil)
	}

	// 获取用户信息
	user := models.NewUser()
	err = user.GetUserByID(c, claims.UserID)
	if err != nil {
		uc.FailAndAbort(c, "获取用户信息失败", err)
	}
	if user.IsEmpty() {
		uc.FailAndAbort(c, "用户不存在", nil, 404)
	}

	// 更新用户头像字段
	user.Avatar = response.Url
	if err := app.DB().WithContext(c).Save(user).Error; err != nil {
		uc.FailAndAbort(c, "更新用户头像失败", err)
	}

	// 返回成功响应
	uc.Success(c, gin.H{
		"url": response.Url,
	})
}

// UpdateBasicInfo 更新当前登录用户基本信息
// @Summary 更新当前登录用户基本信息
// @Description 更新当前登录用户的昵称、性别和描述信息
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param user body models.UpdateBasicInfoRequest true "用户基本信息"
// @Success 200 {object} map[string]interface{} "用户基本信息更新成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /users/updateBasicInfo [put]
// @Security ApiKeyAuth
func (uc *UserController) UpdateBasicInfo(c *gin.Context) {
	var req models.UpdateBasicInfoRequest
	if err := req.Validate(c); err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}

	// 该API只能更新当前用户的账户信息
	currentUserID := common.GetCurrentUserID(c)

	// 检查用户是否存在
	user := models.NewUser()
	err := user.GetUserByID(c, currentUserID)
	if err != nil {
		uc.FailAndAbort(c, err.Error(), err)
	}
	if user.IsEmpty() {
		uc.FailAndAbort(c, "用户不存在", nil, 404)
	}

	// 更新用户基本信息
	user.NickName = req.NickName
	user.Sex = req.Sex
	user.Description = req.Description

	if err := app.DB().WithContext(c).Save(user).Error; err != nil {
		uc.FailAndAbort(c, "更新用户基本信息失败", err)
	}

	uc.SuccessWithMessage(c, "基本信息更新成功", nil)
}

// SwitchTenant 切换租户
// @Summary 切换租户
// @Description 切换当前登录用户的租户，注销当前token并生成新token
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param tenantId path int true "租户ID"
// @Success 200 {object} map[string]interface{} "成功返回新的访问令牌"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "用户未登录"
// @Failure 403 {object} map[string]interface{} "租户不存在或未启用"
// @Router /users/switchTenant/{tenantId} [post]
// @Security ApiKeyAuth
func (uc *UserController) SwitchTenant(c *gin.Context) {
	// 从上下文中获取用户信息
	claims := common.GetClaims(c)
	if claims == nil {
		uc.FailAndAbort(c, "用户未登录", nil)
		return
	}

	// 从路径参数中获取租户ID
	tenantIdStr := c.Param("tenantId")
	if tenantIdStr == "" {
		uc.FailAndAbort(c, "租户ID不能为空", nil)
		return
	}
	tenantId, err := strconv.Atoi(tenantIdStr)
	if err != nil {
		uc.FailAndAbort(c, "租户ID格式错误", err)
		return
	}

	// 获取用户信息
	user := models.NewUser()
	err = user.GetUserByID(c, claims.UserID)
	if err != nil {
		uc.FailAndAbort(c, "用户查询错误", err)
		return
	}
	if user.IsEmpty() {
		uc.FailAndAbort(c, "用户不存在", nil, 404)
		return
	}

	var tenantID uint
	var tenantCode string

	// 当参数tenantId为0且用户的TenantID也为0时，直接使用tenantID=0和tenantCode=""生成token
	if tenantId == 0 && user.TenantID == 0 {
		tenantID = 0
		tenantCode = ""
	} else {
		// 查询租户信息
		tenant := models.NewTenant()
		err = tenant.FindByID(c, uint(tenantId))
		if err != nil {
			uc.FailAndAbort(c, "查询租户错误", err)
			return
		}
		if tenant.IsEmpty() {
			uc.FailAndAbort(c, "租户不存在", nil, 404)
			return
		}
		if tenant.Status != 1 {
			uc.FailAndAbort(c, "租户未启用", nil)
			return
		}

		// 检查用户是否关联该租户
		if user.TenantID > 0 {
			// 用户有默认租户，需要验证是否关联了目标租户
			userTenant := &models.SysUserTenant{}
			err = userTenant.Find(c, func(d *gorm.DB) *gorm.DB {
				return d.Where("user_id = ? AND tenant_id = ?", user.ID, uint(tenantId))
			})
			if err != nil {
				uc.FailAndAbort(c, "查询用户租户关联错误", err)
				return
			}
			if userTenant.IsEmpty() {
				uc.FailAndAbort(c, "用户未关联该租户", nil)
				return
			}
		}

		tenantID = tenant.ID
		tenantCode = tenant.Code
	}

	// 撤销当前 access token
	tokenString, err := common.GetAccessTokenFromHeader(c)
	if err == nil && tokenString != "" {
		app.TokenService.RevokeTokenWithCache(tokenString)
	}

	// 撤销当前 refresh token
	err = app.TokenService.RevokeRefreshToken(claims.UserID)
	if err != nil {
		uc.FailAndAbort(c, "撤销旧token失败", err)
		return
	}

	// 生成新的token
	user.Password = ""
	newToken, err := app.TokenService.GenerateTokenWithCache(&app.ClaimsUser{
		UserID:     user.ID,
		Username:   user.Username,
		TenantID:   tenantID,
		TenantCode: tenantCode,
	})
	if err != nil {
		uc.FailAndAbort(c, "生成新token失败", err)
		return
	}

	// 生成新的refresh token
	newRefreshToken, err := app.TokenService.GenerateRefreshToken(user.ID, tenantID, tenantCode)
	if err != nil {
		uc.FailAndAbort(c, "生成新refresh token失败", err)
		return
	}

	// 解析token获取过期时间
	newClaims, err := app.TokenService.ParseToken(newToken)
	if err != nil {
		uc.FailAndAbort(c, "解析新token失败", err)
		return
	}

	newRefreshClaims, err := app.TokenService.ParseRefreshToken(newRefreshToken)
	if err != nil {
		uc.FailAndAbort(c, "解析新refresh token失败", err)
		return
	}

	uc.Success(c, gin.H{
		"accessToken":         newToken,
		"accessTokenExpires":  newClaims.ExpiresAt.Unix(),
		"refreshToken":        newRefreshToken,
		"refreshTokenExpires": newRefreshClaims.ExpiresAt.Unix(),
	})
}